# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server
components:
  securitySchemes:
    securitySchemes:
      bearerAuth: # arbitrary name for the security scheme
    type: http
    scheme: bearer
    bearerFormat: JWT
...
paths:
...
  /addNewProject:
    post:
      tags:
        - Projects
      summary: Add a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectName:
                  type: string
                projectMembers:
                  type: array
                  items:
                    type: object
                orgID:
                  type: string
      responses:
        '200':
          description: Project added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  projectID:
                    type: string
        '500':
          description: Something went wrong with adding a new project
  /getSpecificProject/{projectID}:
    get:
      tags:
        - projects
      summary: Get specific project details
      parameters:
        - in: path
          name: projectID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  Title:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        projectName:
                          type: string
                        projectStateIDs:
                          type: array
                          items:
                            type: string
                        orgID:
                          type: string
                        projectStates:
                          type: array
                          items:
                            type: object
                            properties:
                              stateName:
                                type: string
                              stateID:
                                type: string
                              position:
                                type: integer
                        membersInfo:
                          type: array
                          items:
                            type: object
                            properties:
                              email:
                                type: string
                              fName:
                                type: string
                              lName:
                                type: string
                              color:
                                type: string
        '500':
          description: Server error
  /getProjects/{orgID}:
    get:
      tags:
        - projects
      summary: Get projects associated with the organization ID
      parameters:
        - in: path
          name: orgID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Projects found
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    type: array
                    items:
                      type: object
                      properties:
                        projectName:
                          type: string
                        projectStateIDs:
                          type: array
                          items:
                            type: string
                        orgID:
                          type: string
                        membersInfo:
                          type: array
                          items:
                            type: object
                            properties:
                              email:
                                type: string
                              fName:
                                type: string
                              lName:
                                type: string
                              color:
                                type: string
        '500':
          description: Internal Server Error
  /updateSingleProjectBoard:
    post:
      tags:
        - projects
      summary: Update single project board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskArray:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Board successfully updated
        '500':
          description: Server error
  /getSingleProject:
    post:
      tags:
        - projects
      summary: Get single project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectID:
                  type: string
      responses:
        '200':
          description: Data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  Title:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        projectName:
                          type: string
                        projectStateIDs:
                          type: array
                          items:
                            type: string
                        orgID:
                          type: string
                        projectStates:
                          type: array
                          items:
                            type: object
                            properties:
                              stateName:
                                type: string
                              stateID:
                                type: string
                              position:
                                type: integer
                        membersInfo:
                          type: array
                          items:
                            type: object
                            properties:
                              email:
                                type: string
                              fName:
                                type: string
                              lName:
                                type: string
                              color:
                                type: string
        '500':
          description: Server error
  /updateProjectData:
    post:
      tags:
        - projects
      summary: Update project data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newProjectName:
                  type: string
                projectID:
                  type: string
                newBoards:
                  type: array
                  items:
                    type: object
                newMembers:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Board successfully updated
        '500':
          description: Server error
  /deleteProject:
    delete:
      tags:
        - projects
      summary: Delete project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectID:
                  type: string
      responses:
        '200':
          description: Deleted project
        '500':
          description: Internal server error
  /updateStatesPos:
    post:
      tags:
        - projects
      summary: Update states position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  ID:
                    type: string
                  position:
                    type: integer
      responses:
        '200':
          description: States updated successfully
        '500':
          description: Internal server error
...